+++ src/test/java/com/CRM_7/stepDefinitions/CompanyStructureStepDefs.java
@@ -7,8 +7,8 @@ import org.junit.Assert;
 public class CompanyStructureStepDefs {
     @Then("the user is able to display the Company Structure")
     public void the_user_is_able_to_display_the_Company_Structure() {
-        EmployeesPage employeesPage = new EmployeesPage();
 
+        EmployeesPage employeesPage = new EmployeesPage();
         Assert.assertTrue(employeesPage.CompanyStructure.isDisplayed());
     }
