+++ src/test/resources/features/@ZULL-808.feature
@@ -1,5 +1,6 @@
-Feature: 
-
+Feature:
+	Background:
+		Given  User logs in with valid credentials
 	#*User Story:* As a user, I should be able to use functions under the Employee menu
 	#
 	# 
@@ -9,7 +10,7 @@ Feature:
 	@ZULL-808
 	Scenario: User should be able to display the Company Structure
 
-
+		And the user clicks on the employees
 		Then the user is able to display the Company Structure
